+++ src/main/java/com/team2/docgram/dto/UserDto.java
@@ -24,9 +24,8 @@ public class UserDto {
 	private String email;
 	private String starmark;
 	private String mytag;
-	private int team;
+	private Long team;
 	private String rank;
 	private String description;
 	private String deptMark;
-	
 }
